+++ commit_msg.txt
@@ -1,0 +1,12 @@
+feat: enforce cool-off period before manual market unpause
+
+Adds a 1-hour minimum cool-off (MIN_UNPAUSE_COOLOFF_SECONDS) that
+must elapse after pause_market() before resume_market() can be
+called by an admin. Prevents rapid pause/unpause cycling.
+
+- New Error::CooloffActive (419)
+- checked_sub used for elapsed-time calculation (overflow-safe)
+- auto_resume_on_expiry() unaffected — cool-off only gates the
+  manual admin resume path
+- Tests: cool-off enforcement, boundary condition, regression
+  coverage for existing InvalidState/Unauthorized paths
